reverting from preact approach
